+++ components/Slider/examples/SliderAutolayoutExampleViewController.h
@@ -14,17 +14,8 @@
  limitations under the License.
  */
 
-#import "AppDelegate.h"
+#import <UIKit/UIKit.h>
 
-@interface AppDelegate ()
-
-@end
-
-@implementation AppDelegate
-
-- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-  // Override point for customization after application launch.
-  return YES;
-}
+@interface SliderAutolayoutExampleViewController : UIViewController
 
 @end